requirements.txt
streamlit>=1.40.0
pandas>=2.2.0
pypdf>=5.0.0
pymupdf>=1.24.0
pytesseract>=0.3.10
Pillow>=10.4.0
python-docx>=1.1.2
